Переписываем код в main,js; рефакторинг
