+++ src/utils/functionsWithDayjs.js
@@ -1,0 +1,22 @@
+import dayjs from 'dayjs';
+import { getRandomInteger } from './functions';
+
+export const generateBeginEndDates = () => {
+  const maxGap = 10;
+  const startDate = dayjs()
+    .add(getRandomInteger(-maxGap, maxGap), 'day')
+    .add(getRandomInteger(-maxGap, maxGap), 'hour')
+    .add(getRandomInteger(-maxGap, maxGap), 'minute');
+  const endDate = startDate
+    .clone()
+    .add(getRandomInteger(0, 14), 'day')
+    .add(getRandomInteger(0, 59), 'hour')
+    .add(getRandomInteger(0, 59), 'minute');
+
+  return {
+    start: startDate.toDate(),
+    end: endDate.toDate()
+  };
+};
+
+export const dateRend = (date, format) => dayjs(date).format(format);
