+++ src/utils/waypointTypes.js
